inference_api/requirements.txt
torch==1.13.0
torchvision==0.14.0
Pillow
